+++ src/main/java/org/markdownwriterfx/editor/MarkdownEditorPane.java
@@ -92,7 +92,7 @@ public class MarkdownEditorPane
 	private String lineSeparator = getLineSeparatorOrDefault();
 
 	public MarkdownEditorPane() {
-		textArea = new StyleClassedTextArea(false);
+		textArea = new MyStyleClassedTextArea(false);
 		textArea.setWrapText(true);
 		textArea.getStyleClass().add("markdown-editor");
 		textArea.getStylesheets().add("org/markdownwriterfx/editor/MarkdownEditor.css");
@@ -348,6 +348,22 @@ public class MarkdownEditorPane
 		textArea.getUndoManager().redo();
 	}
 
+	public void cut() {
+		textArea.cut();
+	}
+
+	public void copy() {
+		textArea.copy();
+	}
+
+	public void paste() {
+		textArea.paste();
+	}
+
+	public void selectAll() {
+		textArea.selectAll();
+	}
+
 	//---- context menu -------------------------------------------------------
 
 	private void showContextMenu(ContextMenuEvent e) {
@@ -378,10 +394,12 @@ public class MarkdownEditorPane
 	}
 
 	private void initContextMenu() {
+		SmartEditActions.initContextMenu(this, contextMenu);
 		spellChecker.initContextMenu(contextMenu);
 	}
 
 	private void updateContextMenu(int characterIndex, int insertionIndex) {
+		SmartEditActions.updateContextMenu(this, contextMenu, characterIndex);
 		spellChecker.updateContextMenu(contextMenu, characterIndex);
 	}
 
@@ -405,5 +423,40 @@ public class MarkdownEditorPane
 			findReplacePane.findNext();
 		else
 			findReplacePane.findPrevious();
+	}
+
+	//---- class MyStyleClassedTextArea ---------------------------------------
+
+	private static class MyStyleClassedTextArea
+		extends StyleClassedTextArea
+	{
+		public MyStyleClassedTextArea(boolean preserveStyle) {
+			super(preserveStyle);
+		}
+
+		@Override
+		public void cut() {
+			selectLineIfEmpty();
+			super.cut();
+		}
+
+		@Override
+		public void copy() {
+			IndexRange oldSelection = selectLineIfEmpty();
+			super.copy();
+			if (oldSelection != null)
+				selectRange(oldSelection.getStart(), oldSelection.getEnd());
+		}
+
+
+		private IndexRange selectLineIfEmpty() {
+			IndexRange oldSelection = null;
+			if (getSelectedText().isEmpty()) {
+				oldSelection = getSelection();
+				selectLine();
+				nextChar(SelectionPolicy.ADJUST);
+			}
+			return oldSelection;
+		}
 	}
 }